Move _ast_to_kast to the kast module
