+++ pyk/src/pyk/kast/__init__.py
@@ -1,3 +1,4 @@
+from ._ast_to_kast import _ast_to_kast
 from .att import EMPTY_ATT, AttEntry, AttKey, Atts, KAtt, WithKAtt
 from .inner import KInner
 from .kast import KAst, kast_term
