+++ pyk/src/pyk/konvert/__init__.py
@@ -1,4 +1,3 @@
-from ._ast_to_kast import _ast_to_kast
 from ._kast_to_kore import _kast_to_kore, kast_to_kore, kflatmodule_to_kore, krule_to_kore
 from ._kore_to_kast import _kore_to_kast, kore_to_kast
 from ._module_to_kore import module_to_kore
